+++ TP1/main.c
@@ -4,8 +4,8 @@
 
 int main(){
     int opcion;
-    int primerNumero = 0;
-    int segundoNumero = 0;
+    float primerNumero = 0;
+    float segundoNumero = 0;
     imprimirMenu(primerNumero,segundoNumero);  //inicializo el menu con los 2 numeros igual a cero
     scanf("%d", &opcion);
     realizarOperacion(opcion,primerNumero,segundoNumero);
